#10399: Add falcon40b demo test with token matching
